UpdateChecker: Split the TextProgressBar from UpdateCheckerForm
